Include namespace in datacontract/datamember attributes Genrate code for visual basic project Refactoring
git-svn-id: https://xsd2code.svn.codeplex.com/svn@29675 4ef1bfeb-0a34-4d02-9305-240e5b07a168

+++ Source/Xsd2Code.SetupLibrary/Xsd2CodeInstaller.cs
@@ -1,34 +1,46 @@
-﻿namespace Xsd2Code.SetupLibrary
-{
-    using System;
-    using System.IO;
-    using System.Diagnostics;
-    using System.Collections;
-    using System.ComponentModel;
-    using System.Configuration.Install;
-    using System.Xml;
+﻿using System;
+using System.Collections;
+using System.ComponentModel;
+using System.Configuration.Install;
+using System.Diagnostics;
+using System.IO;
+using System.Reflection;
+using System.Xml;
 
+namespace Xsd2Code.SetupLibrary
+{
     /// <summary>
     /// Custom action for add-in deployment.
     /// </summary>
-
     [RunInstaller(true)]
     public partial class Xsd2CodeInstaller : Installer
     {
         /// <summary>
         /// Namespace used in the .addin configuration file.
         /// </summary>         
+        private const string ExtNameSpace = "http://schemas.microsoft.com/AutomationExtensibility";
 
-        private const string ExtNameSpace =
-          "http://schemas.microsoft.com/AutomationExtensibility";
+        /// <summary>
+        /// Addin control file name  
+        /// </summary>
+        private const string addinControlFileName = "Xsd2Code.Addin.Addin";
+
+        /// <summary>
+        /// Addin assembly file name  
+        /// </summary>
+        private const string addinAssemblyFileName = "Xsd2Code.Addin.dll";
+
+        /// <summary>
+        /// Saved state key  
+        /// </summary>
+        private const string savedStateKey = "AddinPath";
 
         /// <summary>
         /// Constructor. Initializes components.
         /// </summary>
         public Xsd2CodeInstaller()
-            : base()
         {
-            InitializeComponent();
+            this.InitializeComponent();
         }
 
 
@@ -37,61 +49,56 @@
         /// which will be executed during install process.
         /// </summary>
         /// <param name="savedState">The saved state.</param>
-
         public override void Install(IDictionary savedState)
         {
             base.Install(savedState);
 
             // Parameters required to pass in from installer
+
+/* RU20090225: Not being used. Remove?
             string productName = this.Context.Parameters["ProductName"];
             string assemblyName = this.Context.Parameters["AssemblyName"];
+*/
 
             // Setup .addin path and assembly path
-            string addinTargetPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments), @"Visual Studio 2008\Addins");
-            string assemblyPath = Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location);
-            string addinControlFileName = "Xsd2Code.Addin" + ".Addin";
-            string addinAssemblyFileName = "Xsd2Code.Addin" + ".dll";
+            string addinTargetPath = Path.Combine(
+                
+                Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
+                @"Visual Studio 2008\Addins");
+
+            string assemblyPath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
 
             try
             {
-                DirectoryInfo dirInfo = new DirectoryInfo(addinTargetPath);
-                if (!dirInfo.Exists)
-                {
-                    dirInfo.Create();
-                }
+                var targetFolder = new DirectoryInfo(addinTargetPath);
+                if (!targetFolder.Exists) targetFolder.Create();
 
                 string sourceFile = Path.Combine(assemblyPath, addinControlFileName);
-                XmlDocument doc = new XmlDocument();
-                doc.Load(sourceFile);
-                XmlNamespaceManager xnm = new XmlNamespaceManager(doc.NameTable);
-                xnm.AddNamespace("def", ExtNameSpace);
+
+                var addinXml = new XmlDocument();
+                addinXml.Load(sourceFile);
+
+                var nsmgr = new XmlNamespaceManager(addinXml.NameTable);
+                nsmgr.AddNamespace("def", ExtNameSpace);
 
                 // Update Addin/Assembly node
-
-                XmlNode node = doc.SelectSingleNode("/def:" +
-                    "Extensibility/def:Addin/def:Assembly", xnm);
-                if (node != null)
-                {
-                    node.InnerText = Path.Combine(assemblyPath, addinAssemblyFileName);
-                }
+                SetNodeValue(addinXml, nsmgr,
+                             "/def:Extensibility/def:Addin/def:Assembly",
+                             Path.Combine(assemblyPath, addinAssemblyFileName));
 
                 // Update ToolsOptionsPage/Assembly node
+                SetNodeValue(addinXml, nsmgr,
+                             "/def:Extensibility/def:ToolsOptionsPage/def:Category/def:SubCategory/def:Assembly",
+                             Path.Combine(assemblyPath, addinAssemblyFileName));
 
-                node = doc.SelectSingleNode("/def:Extensibility/def:" +
-                       "ToolsOptionsPage/def:Category/def:SubCategory/def:Assembly", xnm);
-                if (node != null)
-                {
-                    node.InnerText = Path.Combine(assemblyPath, addinAssemblyFileName);
-                }
-
-                doc.Save(sourceFile);
+                addinXml.Save(sourceFile);
 
                 string targetFile = Path.Combine(addinTargetPath, addinControlFileName);
                 File.Copy(sourceFile, targetFile, true);
 
                 // Save AddinPath to be used in Uninstall or Rollback
 
-                savedState.Add("AddinPath", targetFile);
+                savedState.Add(savedStateKey, targetFile);
             }
             catch (Exception ex)
             {
@@ -99,12 +106,16 @@
             }
         }
 
+        private static void SetNodeValue(XmlNode sourceNode, XmlNamespaceManager nsmgr, string xpath, string value)
+        {
+            var node = sourceNode.SelectSingleNode(xpath, nsmgr);
+            if (node != null) node.InnerText = value;
+        }
+
         /// <summary>
         /// Overrides Installer.Rollback, which will be executed during rollback process.
         /// </summary>
-
         /// <param name="savedState">The saved state.</param>
-
         public override void Rollback(IDictionary savedState)
         {
             ////Debugger.Break();
@@ -113,11 +124,8 @@
 
             try
             {
-                string fileName = (string)savedState["AddinPath"];
-                if (File.Exists(fileName))
-                {
-                    File.Delete(fileName);
-                }
+                var fileName = (string)savedState[savedStateKey];
+                if (File.Exists(fileName)) File.Delete(fileName);
             }
             catch (Exception ex)
             {
@@ -128,9 +136,7 @@
         /// <summary>
         /// Overrides Installer.Uninstall, which will be executed during uninstall process.
         /// </summary>
-
         /// <param name="savedState">The saved state.</param>
-
         public override void Uninstall(IDictionary savedState)
         {
             ////Debugger.Break();
@@ -139,11 +145,8 @@
 
             try
             {
-                string fileName = (string)savedState["AddinPath"];
-                if (File.Exists(fileName))
-                {
-                    File.Delete(fileName);
-                }
+                var fileName = (string)savedState[savedStateKey];
+                if (File.Exists(fileName)) File.Delete(fileName);
             }
             catch (Exception ex)
             {
